pyproject.toml: change way to get version
In order to save space, we use shallow clone way to get the repo, but
setuptools-scm not work well with shallow. Eg: with a shallow clone of
v2.2, setuptools-scm will warning like this: '"{wd.path}" is shallow
and may cause error', and a wrong version 0.1.dev1+g3e0db56 is returned.

Change to add a static version to fix above issue

Signed-off-by: Changqing Li <changqing.li@windriver.com>